End check for testing
Bitte geben Sie eine Commit-Beschreibung für Ihre Änderungen ein. Zeilen,

+++ demos/bachelor/FMIndextree_backtrack.cpp
@@ -135,7 +135,7 @@ int main(int argc, char *argv[])
 	
 	unsigned tp = 0;
 	unsigned fn = 0;
-	unsigned fp = 0;
+	unsigned long fp = 0;
 	
 	tim = sysTime();
