+++ demos/bachelor/SAIndextree_backtrack.cpp
@@ -124,7 +124,7 @@ int main(int argc, char *argv[])
 	//SEARCHING
 	unsigned tp = 0;
 	unsigned fn = 0;
-	unsigned fp = 0;
+	unsigned long fp = 0;
 	
 	tim = sysTime();
